+++ R/type-data-frame.R
@@ -103,7 +103,7 @@ vec_cast.data.frame <- function(x, to, ...) {
 #' @export
 #' @method vec_cast.data.frame data.frame
 vec_cast.data.frame.data.frame <- function(x, to, ..., x_arg = "x", to_arg = "to") {
-  .Call(vctrs_df_cast, x, to, x_arg, to_arg)
+  df_cast(x, to, x_arg = x_arg, to_arg = to_arg)
 }
 #' @export
 #' @method vec_cast.data.frame list
@@ -114,6 +114,9 @@ vec_cast.data.frame.list <- function(x, to, ..., x_arg = "x", to_arg = "to") {
 #' @method vec_cast.data.frame default
 vec_cast.data.frame.default <- function(x, to, ..., x_arg = "x", to_arg = "to") {
   vec_default_cast(x, to, x_arg = x_arg, to_arg = to_arg)
+}
+df_cast <- function(x, to, ..., x_arg = "x", to_arg = "to") {
+  .Call(vctrs_df_cast, x, to, x_arg, to_arg)
 }
 
 #' @export